requirements.txt
flask
flask_login
python-dotenv
requests
flask_sqlalchemy
pymysql
numpy